Expand apostrophe section in writing-policy.md
Rewrites the apostrophe section to be more inclusive of when apostrophes are used (and not used) in English. The previous version is worded a little strangely in places and doesn't mention the (in my view) most common error, which is using apostrophes with plurals.

I didn't want to change any subjective policy here, but I think, if you want this to be part of the policy, adding a note that contractions should generally be avoided would not be out of place here. That's a pretty common guideline for formal writing. But for the sake of this PR I just wanted to keep it to objective English grammar stuff.

Apologies that I've been bugging you with this section lately, with me editing the wiki, then that had to get reverted, then I dm'd the QA team with a now-broken link... This should be the last you hear of it from me!