Update PPSSPP developer setup documentation (#129)
* Update emulator-setup-for-developers.md

Updating documentation for PPSSPP setup as the latest public release now supports RAIntegration for developers. It also seems to have a DLL rename.

* Update ppsspp-raintegration-folder.png

This changes the visual filename that the reader sees when attempting to set up PPSSPP with RA_Integration.dll so that it matches the written description provided.

* Update descriptions for clarity

Also removes section for portable install so that the whole thing flows a bit better.