Add further detail to uses of Pause If and Measured If with Measured (#255)
* Add further detail to uses of Pause If and Measured If with Measured

Differentiating between whether to use Pause If or Measured If comes up a fair bit in jr-dev chat.  A lot of times it is assumed you'd only use Measured If. Since there is a behavioral difference between the two where they each have unique use cases, this PR adds some additional detail and organization to the scope-limiting section to make it more visible and hopefully guide developers more effectively in choosing which to use.