DevQuest miscellaneous updates (#290)
* Update devquests.md

General clean up and unifying of the rules.  First version to share with the team, will have updates.

* Update devquests.md

V1.1

* Update devquests.md

V1.2

* Update devquests.md

v1.3 to vote on before merge

* Update devquest-team.md

Update for the team docs for DQ4 to have info on how to handle since any member will be able to do it now

* Update devquests.md

Adding back the mention for tickets resolved for other maintenance quests not being allowed in DQ1